Prefer using docker-image-sizes locally
Due to DockerHub rate limiting

+++ docker/cli-tool-aliases.zsh
@@ -1,8 +1,15 @@
 # Those could also be managed by bin: https://github.com/marcosnils/bin
 
-# cant use gcloud docker auth
-alias docker-image-sizes='docker run --rm -e DIS_IMPL schnatterer/docker-image-size'
-alias docker-image-size='docker run --rm --entrypoint docker-image-size-curl.sh schnatterer/docker-image-size'
+if command -v docker-image-sizes.sh >/dev/null 2>&1; then
+  # Preferred because of DockerHub rate limiting
+  alias docker-image-size="$(which docker-image-sizes.sh)"
+  alias docker-image-sizes="$(which docker-image-size-docker.sh)"
+else 
+  # Use Docker if not on PATH
+  # cant use gcloud docker auth
+  alias docker-image-size='docker run --rm --entrypoint docker-image-size-curl.sh schnatterer/docker-image-size'
+  alias docker-image-sizes='docker run --rm -e DIS_IMPL schnatterer/docker-image-size'
+fi
 
 #alias reg='docker run --rm -v ${HOME}:/root r.j3ss.co/reg'
 #alias terraform='drrmitv ${HOME}:${HOME} --network=host -u "$(id -u):$(id -g)" -w $(pwd) hashicorp/terraform:0.12.24'
